+++ alpha/ghx-engine/src/components/curve_sampler.rs
@@ -13,12 +13,12 @@
 ///
 /// # Returns
 ///
-/// A tuple containing two new `Vec<[f64; 3]>`s, representing the resampled
+/// A tuple containing two new `Vec<[f32; 3]>`s, representing the resampled
 /// polylines with an equal number of points.
 pub fn resample_polylines(
-    polyline_a: &[[f64; 3]],
-    polyline_b: &[[f64; 3]],
-) -> (Vec<[f64; 3]>, Vec<[f64; 3]>) {
+    polyline_a: &[[f32; 3]],
+    polyline_b: &[[f32; 3]],
+) -> (Vec<[f32; 3]>, Vec<[f32; 3]>) {
     if polyline_a.is_empty() || polyline_b.is_empty() {
         return (polyline_a.to_vec(), polyline_b.to_vec());
     }
@@ -44,11 +44,11 @@ pub fn resample_polylines(
     }
 
     let mut resampled_shorter = Vec::with_capacity(longer.len());
-    let shorter_len = (shorter.len() - 1) as f64;
-    let longer_len = (longer.len() - 1) as f64;
+    let shorter_len = (shorter.len() - 1) as f32;
+    let longer_len = (longer.len() - 1) as f32;
 
     for i in 0..longer.len() {
-        let t = i as f64 / longer_len; // Normalized position [0, 1] along the longer polyline
+        let t = i as f32 / longer_len; // Normalized position [0, 1] along the longer polyline
         let shorter_t = t * shorter_len; // Corresponding position along the shorter polyline
         let index = shorter_t.floor() as usize;
         let frac = shorter_t.fract();
